Add "Add to Calendar" link to concert page
- Consolidate ICS logic into lib/calendar.ts
- Add "Add to Calendar" thing to Concert page that downloads an ICS file of this concert for you to use on calendar app
